Refactoring: separate HydraAuctionUtils.Monads.Actors
