+++ hydra-auction.cabal
@@ -259,6 +259,7 @@ library hydra-auction-utils
     HydraAuctionUtils.Extras.PlutusOrphans
     HydraAuctionUtils.Fixture
     HydraAuctionUtils.Monads
+    HydraAuctionUtils.Monads.Actors
     HydraAuctionUtils.Plutus
     HydraAuctionUtils.Tracing
     HydraAuctionUtils.Tx.AutoCreateTx
